Refactor DevsktopMain and enhance documentation for window hooks
- Removed unused AboutSolvedokuWindow and AboutRPSSketchWindow components from DevsktopMain, streamlining the interface.
- Updated index.ts to include new Tailwind color utilities, improving accessibility for styling.
- Added comprehensive JSDoc comments to various window hooks, including useAboutSolvedokuWindow, useAboutThisSiteWindow, useMainLandingWindow, and useSolvedokuWindow, enhancing clarity on their functionality and return values.
- Improved the organization of imports across multiple files for better readability and maintainability.
